+++ simple_history/tests/admin.py
@@ -9,6 +9,7 @@ from .models import (
     ConcreteExternal,
     Document,
     Employee,
+    FileModel,
     Paper,
     Person,
     Poll
@@ -24,6 +25,14 @@ class ChoiceAdmin(SimpleHistoryAdmin):
     history_list_display = ['votes']
 
 
+class FileModelAdmin(SimpleHistoryAdmin):
+
+    def test_method(self, obj):
+        return 'test_method_value'
+
+    history_list_display = ['title', 'test_method']
+
+
 admin.site.register(Poll, SimpleHistoryAdmin)
 admin.site.register(Choice, ChoiceAdmin)
 admin.site.register(Person, PersonAdmin)
@@ -32,3 +41,4 @@ admin.site.register(Document, SimpleHistoryAdmin)
 admin.site.register(Paper, SimpleHistoryAdmin)
 admin.site.register(Employee, SimpleHistoryAdmin)
 admin.site.register(ConcreteExternal, SimpleHistoryAdmin)
+admin.site.register(FileModel, FileModelAdmin)