+++ simple_history/tests/models.py
@@ -119,6 +119,7 @@ class Person(models.Model):
 
 
 class FileModel(models.Model):
+    title = models.CharField(max_length=100)
     file = models.FileField(upload_to='files')
     history = HistoricalRecords()